requirements.txt


# Core Deep Learning
torch==2.1.0
torchvision==0.16.0
torchmetrics==1.1.1

# Image Processing & Augmentation
opencv-python==4.8.0.76
Pillow==10.0.0

# Data Handling & I/O
numpy==1.24.4
pandas==2.1.0
scikit-learn==1.3.0
h5py==3.9.0

# Visualization & Debugging
matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.14.1
tqdm==4.66.1
snakeviz==2.2.0

# Developer Tools & CLI
typer==0.9.0
ruff==0.4.1


